requirements.txt
Django==3.1
psycopg2-binary
djangorestframework
django-rest-auth
django-allauth
Pillow
celery
drf-yasg